now possible to run prediction with bpl-next, with some errors
- Hangs  if running multithreaded, possibly because bpl-next  uses jax which is itself uses multiprocessing. See https://github.com/google/jax/issues/1805 but changing multiprocessing in AIrsenal gives errors about sqlalchemy session not being pickle-able.

- bpl-next predictions occasionally have nan values
